# esto-laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ESTO Plateforme Doctorale API",
        "description": "API REST complète de la Plateforme de Gestion Doctorale — École Supérieure de Technologie d'Oujda (UMP)",
        "contact": {
            "email": "admin@ump.ac.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Lister tous les utilisateurs",
                "operationId": "3265ab79c9b880008ed804d6b5e3059a",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filtrer par rôle (admin, directeur, professeur, candidat)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lab_id",
                        "in": "query",
                        "description": "Filtrer par laboratoire",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un nouvel utilisateur (Admin)",
                "operationId": "7cfe4acccb16b697b910fb2c3d8eb350",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Benali"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Youssef"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "y.benali@esto.ma"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password123!"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "directeur",
                                            "professeur",
                                            "candidat"
                                        ],
                                        "example": "professeur"
                                    },
                                    "laboratoire_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Modifier un utilisateur",
                "operationId": "d273e38b4ab89cda9bf08737d4bd1dcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "directeur",
                                            "professeur",
                                            "candidat"
                                        ]
                                    },
                                    "laboratoire_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "60aeabff552065f790abfd07c24a27ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Statistiques globales du tableau de bord Admin",
                "operationId": "b1b20f78f84973a48b4e9c380957babe",
                "responses": {
                    "200": {
                        "description": "Statistiques: nombre d'utilisateurs, labs, sujets, candidatures par statut"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}